Move in memory connector test Add traces indicating the current config.
